Fine tuning of image position via frontmatter
- image_anchor
- image_position
- image_position_article
- image_position_list

By default, images are cropped using [Smartcrop](https://gohugo.io/content-management/image-processing/#smart-cropping-of-images), which sometimes gets it wrong.  In those cases, the anchor may be overridden for the feature image in the page frontmatter, e.g.

```
image_anchor = "Center"
```

Permitted values for the anchor are as described in the [Hugo documentation](https://gohugo.io/content-management/image-processing/#anchor).

The [PR for this feature](https://github.com/CaiJimmy/hugo-theme-stack/pull/417) was rejected, in order to keep the theme simple.

This styles the image in the main list view using CSS `object-position`.  For example

```
image_position = "50% 20%"
```

It is completely independent of `image_anchor`.

There are also `image_position_article` and `image_position_list` for applying the position to the main article view and to the list view independently.